Add script for generating correct bignum RF label indices
It turns out that our synthesis flow results in the following
correspondence between the rf indices and the actual WDRs:

  w0  - rf[9983:9672]
  w1  - rf[9671:9360]
  w2  - rf[9359:9048]
  .   .        .
  w31 - rf[ 311:   0]

As a result, its pretty tedious to compute the actual indices. This
is automated by the newly added script. In addition it supports a width
and offset argument to reduce the verification time by only defining a
subset of the register content as secret.

In the input file, the user can specify e.g.

  w6: secret 0
  w7: secret 0

which means that WDR6 and WDR7 will contain the shares for secret 0.

The script can be used e.g. with

  ./examples/otbn/labels/generate_bignum_rf_labels.py \
      -i examples/otbn/labels/isw_and_labels.txt \
      -o tmp/labels.txt \
      -w 1 -s 0

If the input file contains other labels than WDR labels, the script just
will just forward these untouched. This allows to use the convenient
format for WDR labels and the existing format for other labels.