Modify boolean_to_arithmetic routine against SCA
This changes are done regarding to outcomes of the COCO-Alma
tool usage. Moreover, many more rules are put in the OTBN Security
Guidance Document. Please see it before developing algorithms
on OTBN which handle secret data.

If we don't have these changes, the tool reports there
are transient leakage and we may see leakage on the final
product even if we don't see a leakage on FPGA experiments.

We are also seeing a stable leakage on the MSB of the
intermediate values. We acknowledge that and since the
algorithm runs with 64-bit excess randomness, we don't
expect that to be possible to use that leakage and
retrieve secret values. We also verified that the stable
leakage disappeared after running the routine on 320-bit
instead of 321-bit.